fix: unit test needs to be adapted
Previously, we were gracefully handling this case, but honestly it is
not the best approach to pretend there's an empty structure if the server
breaks the API and returns `"null"` rather than an object.

That said, it was still awesome to have this test in place because it
helped us to figure out this extra condition of which httpclientx should
be aware and that this problem needs to be dealt with systematically
inside the httpclientx package.