feat(testingx): model OONI backend register/login flows (#1567)
This diff introduces an `http.Handler` that models the register and
login flows of the OONI backend. This diff will allow us to write better
tests for register, login, and logged APIs. This work is part of
https://github.com/ooni/probe/issues/2718.

While there, commit the `internal/urlx` package that is part of
https://github.com/ooni/probe-cli/pull/1560, because it helps us more
easily write code to write tests for this new `testingx` functionality.