feat: introduce the httpclientx package (#1569)
This package tries to replace both `httpx` and `httpapi`. This commit
includes a design document explaining the rationale and the process for
performing this replacement. I have been using this package for a while
at https://github.com/ooni/probe-cli/pull/1560. It seems clear to me
that, at this point, we want to merge. So, I have extracted just the
package from https://github.com/ooni/probe-cli/pull/1560 to reduce the
size of the diff.

The TL;DR of this new package is to ensure we have the same
functionality of `httpx` and `httpapi` but we're also simplifying the
codebase by (a) replacing two other packages and (b) dropping the
requirement that we wanted to generate a swagger from the API
implementation, which lead to overly complex code, and, what is possibly
worst, to code that at the end of the day we were not using for
comparing to the backend's swagger.

Part of https://github.com/ooni/probe/issues/2700.