refactor(webconnectivity*): don't call httpapi directly (#1582)
The https://github.com/ooni/probe-cli/pull/1581 pull request missed that
adding a `CallWebConnectivityTestHelper` method to `*engine.Session`
means we need to mock such a method in `webconnectivityqa`.

I could have mocked the method, but I was not super happy about doing
this, because I would rather have wanted to add more QA tests inside of
`webconnectivityqa` making sure we're indeed falling back when using
test helpers. (We have unit tests testing that, but it would be nice to
have integration tests, so we could observe the overall code behavior.)

My initial approach for allowing this was to create a
`CallWebConnectivityTestHelper` function inside the engine, taking a
`model.ExperimentSession` as its fourth argument and then modifying the
`CallWebConnectivityTestHelper` method to just invoke the function
passing to it the session. However, this approach created import cycles
in tests.

To avoid import cycles, I moved the `CallWebConnectivityTestHelper`
function to `webconnectivityalgo` and modified the mocks inside
`webconnectivityqa` to call such a function. This is where I paused and
realized that the cleanest solution is actually to just always call the
`CallWebConnectivityTestHelper` function in `webconnectivityalgo` and
undo part of the changes in https://github.com/ooni/probe-cli/pull/1581
to avoid having a `CallWebConnectivityTestHelper` method inside of
`model.ExperimentSession`. And so I also implemented this change.

All of this leads us to the current diff, which is part of
https://github.com/ooni/probe/issues/2725.