fix(enginenetx): compute zero time when we start dialing
This diff addresses a bug observed on the wild where a slow DNS
causes several tactics to be ready concurrently.

If we want several tactics to be ready concurrently, we should
arrange for that, and for now BTW that's not the case.

Part of https://github.com/ooni/probe/issues/2704.