fix(enginenetx): compute zero time when we start dialing (#1594)
Previously, the code was computing the zero time when we started
resolving. However, I have observed in the wild that, if the DNS lookup
time is high, we're going to have several ready tactics. We did not
previously see this bug because we gave priority to bridges and stats
tactics, hence we always had some ready tactics from the get go.

This PR is part of settling the dust after the changes requested in the
https://github.com/ooni/probe-cli/pull/1552 review.

The related tracking issue is https://github.com/ooni/probe/issues/2704.

---------

Co-authored-by: Arturo Filastò <arturo@filasto.net>