refactor: MeasurementTarget -> MeasurementInput
While starting to work on richer input, I realised that it would
be very handy to reserve the name "target" to describe a richer input
target as the tuple containing options and some input.

However, using such a naming would clash with the definition of
MeasurementTarget, which is a string definition that is only there
to say semantically that the Input field of a measurement is the
thing that we should really measure.

Therefore, for making semantic space for targets, I am going to
refactor MeasurementTarget to be MeasurementInput.

No functional change.

Part of https://github.com/ooni/probe/issues/2607