cleanup: remove asynchronous experiments (#1613)
We originally developed asynchronous experiments as a mean to support
websteps, a nettest that returned more than one measurement per
invocation of its Measure method.

Since then, we removed websteps. Therefore, this code is currently
technically unused. Additionally, this code further complicates
implementing richer input, because it is another way of performing
measurements.

As such, in the interest of switfly moving forward with richer input
_and_ of simplifying the engine, we are now removing this unused
functionality from the tree.

While there, better the documentation of OnProgress and undeprecate
functions to open/close reports since it's clear that, for now, using a
submitter in cmd/ooniprobe is a bit of a stretch given our current
goals. So, it feels best to avoid deprecating until we have nice plan
for replacement.

To have more confidence that the job we did is ~correct, we use the
following table to cross compare the operations that we previously
performed for running sync experiments (i.e., all experiments) in an
async way to the code we're using after this diff to run experiments.
(Note that the diff itself is such that you can easily see the deleted
and the added code inside of the `experiment.go` file.) In both cases,
we're looking at the operations performed starting from
`MeasureWithContext`.

| Operation                   | Before | After |          
| --------------------------- | ------ | ----- |
| MaybeLookupLocationContext  | yes    | yes   |
| use e.session.byteCounter   | yes    | yes   |
| use e.byteCounter           | yes    | yes   |
| newMeasurement              | yes    | yes   |
| save start time             | yes    | yes   |
| initialize experiment args  | yes    | yes   |
| call measurer.Run           | yes    | yes   |
| save end time               | yes    | yes   |
| compute measurement runtime | yes    | yes   |
| scrub measurement           | yes    | yes   | 
| return measurement          | yes    | yes   |

Part of https://github.com/ooni/probe/issues/2607

Depends on https://github.com/ooni/probe-cli/pull/1612