refactor: construct TargetLoader using ExperimentBuilder (#1617)
This diff completes the set of preliminary richer input diffs. We build
the TargetLoader using the ExperimentBuilder, which in turn uses a
registry.Factory under the hood. This means that we can load targets for
each experiment, because the actual implementation of the TargetLoader
can be experiment dependent.

Part of https://github.com/ooni/probe/issues/2607