refactor(oonimkall): fakeSuccessful{Run => Deps} (#1631)
This diff contains a trivial rename of fakeSuccessfulRun back to
fakeSuccessfulDeps. I made the rename from Deps to Run in a previous
commit, but now I realize it was a mistake.

Because this is a trivial patch generated with a tool, I am going to
self merge it w/o review.

Part of https://github.com/ooni/probe/issues/2752.